warn: use OOUI for "linked article" and other inputs
This involves some restructuring of the code that generates the input field on basis of the selected warning template.

The old setup created the "Linked article" field to begin with, and dynamically modified it as necessary, taking care to tweak the label and tooltip, etc.

The new setup recreates the field from scratch every time the selected template is changed. While being less "efficient" by a few milliseconds, it is a lot more flexible - we can specify different classes/tooltips for inputs corresponding to different templates. It also becomes easy to change or remove OOUI fields.