xfd: use OOUI for selecting TFD/CFD merge targets
Fairly straightforward in case of TFD. But in case of CFD, some restructuring is done so that the inputs are recreated everytime the CFD subtype is changed.

Now, on choosing "Deletion" or "Stub Deletion", no input field as shown (earlier a disabled one used to be shown). A positive side-effect, I'd say.