morebits: date: format per user time zone
A recent MW patch https://gerrit.wikimedia.org/r/c/mediawiki/core/+/679959 fixes the issue with `mw.user.options.get('timecorrection')` not giving the DST-corrected user timezone offset.

This commit adds the ability in morebits to format dates per this timezone and makes it the default, consistent with MediaWiki using it in page histories, logs, diffs, etc. The time shown will still be off by an hour for DST-affected users when talking about time on the other side of the DST switchover. But that is a much less significant bug as the difference of an hour probably doesn't matter when the time is anyway quite far ahead or behind in the future.

Until now, we've been using either system time (few places) or UTC (most places, but variously with and without the "(UTC)" appended). UTC is not ideal because not all wikis use UTC as the wiki default time, and we don't have the local timezone string available. No timezone string is needed when using user time (since MW doesn't use them either in logs/diffs.) Even for wikis where UTC is default, it doesn't make any sense with `.calendar()`. That is, saying `Yesterday at 8:30 PM` is silly when for many users the said time is neither yesterday nor 8:30.